review feedback and check if other servers are allowed
