+++ app/components/channel_list/header/header.tsx
@@ -1,8 +1,9 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React from 'react';
+import React, {useEffect} from 'react';
 import {Text, View} from 'react-native';
+import Animated, {useAnimatedStyle, useSharedValue, withTiming} from 'react-native-reanimated';
 
 import CompassIcon from '@components/compass_icon';
 import TouchableWithFeedback from '@components/touchable_with_feedback';
@@ -24,9 +25,6 @@ const getStyles = makeStyleSheetFromTheme((theme: Theme) => ({
     subHeadingStyles: {
         color: changeOpacity(theme.sidebarText, 0.64),
         ...typography('Heading', 50),
-    },
-    iconPad: {
-        marginLeft: 44,
     },
     headerRow: {
         flexDirection: 'row',
@@ -57,10 +55,18 @@ const getStyles = makeStyleSheetFromTheme((theme: Theme) => ({
 const ChannelListHeader = ({displayName, iconPad}: Props) => {
     const theme = useTheme();
     const serverDisplayName = useServerDisplayName();
+    const marginLeft = useSharedValue(iconPad ? 44 : 0);
     const styles = getStyles(theme);
+    const animatedStyle = useAnimatedStyle(() => ({
+        marginLeft: withTiming(marginLeft.value, {duration: 350}),
+    }), []);
+
+    useEffect(() => {
+        marginLeft.value = iconPad ? 44 : 0;
+    }, [iconPad]);
 
     return (
-        <View style={iconPad && styles.iconPad}>
+        <Animated.View style={animatedStyle}>
             <View style={styles.headerRow}>
                 <View style={styles.headerRow}>
                     <Text style={styles.headingStyles}>
@@ -83,7 +89,7 @@ const ChannelListHeader = ({displayName, iconPad}: Props) => {
             <Text style={styles.subHeadingStyles}>
                 {serverDisplayName}
             </Text>
-        </View>
+        </Animated.View>
     );
 };
