+++ package.json
@@ -139,6 +139,7 @@
     "jest-cli": "27.3.1",
     "jetifier": "2.0.0",
     "metro-react-native-babel-preset": "0.66.2",
+    "mmjstool": "github:mattermost/mattermost-utilities#519b99a4e51e6c67a0dbd46a6efdff27dc835aaa",
     "mock-async-storage": "2.2.0",
     "nock": "13.1.4",
     "patch-package": "6.4.7",
